pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whoop-mcp"
version = "0.1.0"
description = "Whoop fitness data MCP server for Claude"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
whoop-mcp = "whoop_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/whoop_mcp"]
